feat: merge in changes from lonko and NevrDid's N64 pak
These new config changes the controller signficantly, but also increases the performance of the emulator, making gaming buttery smooth.

Since the controller layout is radically different but performance is improved, the performance changes were copied over to the existing default controller config, and the new controller config is offered as an alternative Controller layout.

Closes #8