fix: do not delete game resume slot if the saves were not restored
Due to mucking about with save paths, it is possible to get into a state where the save is not restored if exiting out of the emulator settings menu  and assume that this means there was no save. We now keep track of whether saves were restored before updating the game's resume slot.