refactor: use options specified in config.json to drive minui-list
This allows for faster, more intuitive emulator settings interactions as folks can just scroll back and forth across options without needing to stop/start minui-list to persist settings. The new action button is used to immediately play with the configured settings, while the confirm button is used solely to save the settings for the game.

One nice thing about this is that adding new options is quite a bit simpler, and its easier for users to see what options are available by checking the config.json for the pak.